Implements a mechanism to clear the cache of the examples tree.
This implements a mechanism very similar to WebForms EE that re-reads the
config.xml file (root) so that changes on it won't require either a process,
thread or server restart before they reflect on the server.

By accessing the main view passing the ?clearExamplesCache parameter, the
static variable will be nullified so that it is re-read the next time it is
necessary.

As in WebForms EE, a dialog alert will be displayed as a feedback for the
procedure successfully being executed.