Fixes Draw > Rotate Text example.
The 'change' event in the slider was being called way too early, before
the drawcontainer below was ready to be used, and this was triggering an
error on load.

Just moved the change handler to a function and ensured the function
only runs when the draw container is ready to be manipulated.

This fixes #6 item 52.