Introduces a KeyMap example for the component-level key binding.
The new examples illustrates key mapping at component-level (bound to
the viewport in the case), as well as key binding variations.

In MVC Razor syntax, it also illustrates the X.KeyBindItem integrated,
removing the requirement to define it via lambda expressions, but in a
more natural Razor syntax format, and also the intellisense-friendly
@<Text><script> tag support, as well as token expansion support within.

This covers usage of the new approach implemented for key mapping in
ExtJS 6, that was broken/unusable in Ext.NET since version 4.0.0.

Related github issue: extnet/Ext.NET#1466.