Makes the list of themes dynamic, from Ext.Net.Theme enum.
Now it pulls the names of themes from the enum, so that newly added
themes will reflect to the examples explorer seamlessly.

Exceptions are to the 'None' theme that only makes sense while making
custom themes, and the 'Default' theme that is just displayed as 'Blue'
in the themes list.

This allows for switching the examples to the newly introduced Graphite
theme.