Renames references to Theme.Default and 'blue' to 'Classic'.
The theme has not been the default one for a long long time, and in
Ext.NET 5 its name was changed to reflect the name currently used in Ext
JS: 'classic'.

References to 'blue' as the theme name, as well as the enum reference,
Ext.Net.Theme.Default were changed accordingly.

This comes as a conseuence of extnet/Ext.NET#1602.