Introduces Associations > Lazy Load example.
This is an effort porting the example from WebForms, and uses partial
views instead of custom controls for the lazy-loaded panels within it.

The example currently does not load the order details, it is missing the
_order() method in every client-side customer data entry.

The store needed to force proxy as empty/null because from the partial
view, the store is always coming with proxy type as 'memory' for some
reason.