Adds the Charts > Combination > Binding Tabs example.
This is simply a port from WebForms corresponding example. The ViewModel
in here bound using MVC's ViewData that is synced between the controller
and view to properly fill the tab panel's ViewModel. An usual '@model'
would probably also work just fine, but this shows an alternative
approach to binding initial data to Ext.NET MVC components.