Starts work in the Charts > Combination > Infographics example.
The example is partially implemented, thus its view folder is named
after "_Infographic"; the leading underscore prevents it from being
listed in Ext.NET's shell.

It currently loads its data and shows the main graph elements in the
first of the Polar Charts.

The example consists in two Polar Charts and one Cartesian charts (bar
chart).

Still in the first polar chart, it is missing hover effects and the
custom ArcTextSprite component.